+++ kpm/kpm-frontend/src/app.tsx
@@ -5,6 +5,24 @@ import { Menu } from "./Menu";
 
 import "./app.scss";
 
+const IS_DEV = process.env.NODE_ENV !== "production";
+
+export type TCurrentUser = {
+  kthid: string,
+  display_name: string,
+  email: string,
+  username: string,
+  exp: number,
+}
+
+declare global {
+  interface Window {
+    __kpmCurrentUser__: TCurrentUser;
+  }
+}
+
+export const currentUser: TCurrentUser = window.__kpmCurrentUser__ || (IS_DEV && { kthid: 'u19t0qf2', display_name: 'Sebastian Ware', email: 'jhsware@kth.se', username: 'jhsware', exp: 1668683814 })
+
 type TCreateRouterProps = {
   hasStudies: boolean;
   hasTeaching: boolean;
