+++ kpm/kpm-frontend/src/components/utils.tsx
@@ -1,3 +1,8 @@
 export function linkClassName({ isActive }: { isActive: boolean }) {
   return isActive ? " active" : "";
 }
+
+
+export function formatDisplayName(name: string) {
+  return name.split(" ")[0];
+}
