+++ kpm/kpm-frontend/src/panes/profile.tsx
@@ -1,38 +1,22 @@
 import * as React from "react";
-import { useLoaderData } from "react-router-dom";
-import { ErrorMessage, LoadingPlaceholder } from "../components/common";
-import { MenuPane } from "../components/menu";
-import { createApiUri, useDataFecther } from "./utils";
+import { MenuPane, MenuPaneHeader } from "../components/menu";
+import { createApiUri } from "./utils";
 import { i18n } from "../i18n/i18n";
-
-export async function loaderProfile({ request }: any = {}) {
-  const res = await fetch("/kpm/api", {
-    signal: request?.signal,
-  });
-  const json = await res.json();
-  return json;
-}
+import { currentUser } from "../app";
+import { formatDisplayName } from "../components/utils";
 
 export function Profile() {
-  const { res, loading, error } = useDataFecther<{ msg: string }>(
-    loaderProfile
-  );
-  const { msg } = res || {};
-  // const { msg } = useLoaderData() as { msg: string };
   return (
     <MenuPane>
-      {loading && <LoadingPlaceholder />}
-      {error && <ErrorMessage error={error} />}
-      {msg && (
+      <MenuPaneHeader title={`${formatDisplayName(currentUser.display_name)} - ${i18n("Settings")}`}>
+          <a className="kpm-button" href={createApiUri("/auth/logout")}>{i18n("Logout")}</a>
+      </MenuPaneHeader>
         <h2>
-          {i18n("Profile")} {msg}
+          
         </h2>
-      )}
-      {msg && (
         <p>
-          <a href={createApiUri("/auth/logout")}>{i18n("Logout")}</a>
         </p>
-      )}
     </MenuPane>
   );
 }
+
